==> Collaborative/1.data_precessing.ipynb <==
import pandas as pd
import numpy as np
from collections import defaultdict
import scipy.io as sio
import scipy.sparse as ss
import pickle
from sklearn.model_selection import train_test_split

#读取数据
data = pd.read_csv("triplet_dataset_sub.csv")
data.head()

data.info()

X_data = data.drop(['play_count'], axis=1)
y_data = data['play_count']

#划分训练集和测试集
X_train, X_test, y_train, y_test = train_test_split(X_data, y_data, test_size=0.2)
train_data = pd.concat([X_train, y_train], axis=1)
test_data = pd.concat([X_test, y_test], axis=1)

train_data.head()

test_data.head()

#获取所有用户和所有歌曲的ID标识
total_users_id = train_data['user'].unique()
total_songs_id = train_data['song'].unique()
print('total user {}, total song {}'.format(len(total_users_id), len(total_songs_id)))

#对用户和歌曲ID进行编号
users_index = dict()
songs_index = dict()

for index, user in enumerate(total_users_id):
    users_index[user] = index
    print('user:{}, index:{}'.format(user, index))

for index, song  in enumerate(total_songs_id):
    songs_index[song] = index
    print('song:{}, index:{}'.format(song, index))

#数据保存
pickle.dump(total_users_id, open('total_users_id.pkl', 'wb'))
pickle.dump(total_songs_id, open('total_songs_id.pkl', 'wb'))
pickle.dump(users_index, open('users_index.pkl', 'wb'))
pickle.dump(songs_index, open('songs_index.pkl', 'wb'))

#记录用户对歌曲是否感兴趣，个人认为至少播放2次才是感兴趣
interest = np.array(train_data['play_count'])
interest[interest < 2] = 0
interest[interest >= 2] = 1
train_data['interest'] = interest
train_data.head()

#根据次数转换为分数
min_count = train_data['play_count'].min()
max_count = train_data['play_count'].max()
diff = max_count - min_count
cut_list = [0, min_count+diff//5, min_count+diff*2//5, min_count+diff*3//5, min_count+diff*4//5, min_count+diff]
labels = [1, 2, 3, 4, 5]
train_data['score'] = pd.cut(train_data['play_count'], cut_list, labels=labels)
train_data.head()

#统计用户听过的歌曲，歌曲有哪些用户听过
user_song_dict = defaultdict(set)
song_user_dict = defaultdict(set)

# #记录用户对歌曲是否感兴趣和次数
user_song_interest = ss.dok_matrix((len(total_users_id), len(total_songs_id)))
user_song_score = ss.dok_matrix((len(total_users_id), len(total_songs_id)))

for index, row in train_data.iterrows():
    cur_user_id = users_index[row['user']]
    cur_song_id = songs_index[row['song']]
    
    if(row.interest != 0):
        #记录用户听过的歌曲
        user_song_dict[cur_user_id].add(cur_song_id)

        #记录听过歌曲的用户
        song_user_dict[cur_song_id].add(cur_user_id)

        #记录用户对歌曲的兴趣
        user_song_interest[cur_user_id, cur_song_id] = row['interest']
        user_song_score[cur_user_id, cur_song_id] = row['score']

user_song_score

sio.mmwrite("user_song_interest", user_song_interest)
sio.mmwrite("user_song_score", user_song_interest)

#存储数据
pickle.dump(user_song_dict, open('user_song_dict.pkl', 'wb'))
pickle.dump(song_user_dict, open('song_user_dict.pkl', 'wb'))

#保存测试数据
interest = np.array(test_data['play_count'])
interest[interest < 2] = 0
interest[interest >= 2] = 1
test_data['interest'] = interest

#根据次数转换为分数
min_count = test_data['play_count'].min()
max_count = test_data['play_count'].max()
diff = max_count - min_count
cut_list = [0, min_count+diff//5, min_count+diff*2//5, min_count+diff*3//5, min_count+diff*4//5, min_count+diff]
labels = [1, 2, 3, 4, 5]
test_data['score'] = pd.cut(test_data['play_count'], cut_list, labels=labels)
test_data.head()

test_data.to_csv('test_data.csv', index=True)

train_data.head()

train_data.to_csv("train_data.csv", index=True)
